requirements.txt
django
six
django-recaptcha